@@ -27,38 +27,41 @@
     "@babel/plugin-transform-class-properties": "^7.23.3",
     "@babel/preset-env": "^7.18.2",
     "@babel/preset-react": "^7.17.12",
+    "@tanstack/react-router": "^1.130.12",
     "@types/photoswipe": "^4.1.2",
     "@types/react": "^18.0.14",
     "@wordpress/commands": "^0.16.0",
-    "@wordpress/components": "^25.11.0",
+    "@wordpress/components": "^28.9.0",
+    "@wordpress/dataviews": "^5.0.0",
     "@wordpress/element": "^4.9.0",
     "@wordpress/eslint-plugin": "^12.8.0",
     "@wordpress/i18n": "^4.11.0",
     "@wordpress/icons": "^9.36.0",
-    "@wordpress/interface": "^8.3.0",
-    "@wordpress/scripts": "^30.7.0",
+    "@wordpress/interface": "^6.9.0",
+    "@wordpress/scripts": "^30.20.0",
     "ajv": "^8.17.1",
     "axios": "^1.6.1",
     "babel-plugin-macros": "^3.1.0",
     "classnames": "^2.3.1",
     "css-loader": "^6.7.1",
-    "dompurify": "^2.4.0",
+    "dompurify": "^3.2.5",
     "eslint": "^8.18.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "grunt": "^1.6.1",
     "grunt-contrib-compress": "^2.0.0",
     "lodash.uniqueid": "^4.0.1",
-    "lucide-react": "^0.292.0",
+    "lucide-react": "^0.468.0",
     "mini-css-extract-plugin": "^2.6.1",
     "prettier": "^2.7.1",
     "prettier-eslint": "^15.0.1",
     "process": "^0.11.10",
     "prop-types": "^15.8.1",
     "qs": "^6.10.5",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.48.2",
+    "react-spinners": "^0.17.0",
     "resolve-url-loader": "^5.0.0",
     "sass": "^1.66.1",
     "sass-loader": "^13.0.2",
@@ -71,7 +74,11 @@
   },
   "dependencies": {
     "@fancyapps/ui": "^5.0.36",
+    "@tanstack/react-query": "^5.84.1",
     "@wordpress/a11y": "^3.11.0",
+    "@wordpress/block-editor": "^14.4.0",
+    "@wordpress/block-serialization-default-parser": "^5.9.0",
+    "@wordpress/url": "^4.9.0",
     "grunt-cli": "^1.4.3",
     "html-to-react": "^1.4.8"
   }
