@@ -1,24 +1,23 @@
-<?php return array (
-  'root' => 
-  array (
-    'pretty_version' => 'dev-main',
-    'version' => 'dev-main',
-    'aliases' => 
-    array (
+<?php return array(
+    'root' => array(
+        'pretty_version' => 'dev-main',
+        'version' => 'dev-main',
+        'type' => 'project',
+        'install_path' => __DIR__ . '/../../',
+        'aliases' => array(),
+        'reference' => '603db79fd04055ba07509f1af4a56268850842c0',
+        'name' => 'dlxplugins/pattern-wrangler',
+        'dev' => true,
     ),
-    'reference' => '98fdd65c94cacda5c9c5206215cac2130140a5db',
-    'name' => 'dlxplugins/pattern-wrangler',
-  ),
-  'versions' => 
-  array (
-    'dlxplugins/pattern-wrangler' => 
-    array (
-      'pretty_version' => 'dev-main',
-      'version' => 'dev-main',
-      'aliases' => 
-      array (
-      ),
-      'reference' => '98fdd65c94cacda5c9c5206215cac2130140a5db',
+    'versions' => array(
+        'dlxplugins/pattern-wrangler' => array(
+            'pretty_version' => 'dev-main',
+            'version' => 'dev-main',
+            'type' => 'project',
+            'install_path' => __DIR__ . '/../../',
+            'aliases' => array(),
+            'reference' => '603db79fd04055ba07509f1af4a56268850842c0',
+            'dev_requirement' => false,
+        ),
     ),
-  ),
 );
