@@ -10,7 +10,7 @@ import {
 	Button,
 } from '@wordpress/components';
 import { useAsyncResource } from 'use-async-resource';
-import { AlertTriangle, CheckCircle } from 'lucide-react';
+import { Info, AlertTriangle } from 'lucide-react';
 
 import { __, _n } from '@wordpress/i18n';
 import { useForm, Controller, useWatch, useFormState } from 'react-hook-form';
@@ -20,30 +20,6 @@ import classNames from 'classnames';
 import SendCommand from '../../utils/SendCommand';
 import Notice from '../../components/Notice';
 import SaveResetButtons from '../../components/SaveResetButtons';
-
-const retrieveOptions = () => {
-	return SendCommand( 'dlx_pw_get_options', {
-		nonce: dlxPatternWranglerAdmin.getNonce,
-	} );
-};
-
-const Main = ( props ) => {
-	const [ defaults ] = useAsyncResource(
-		retrieveOptions,
-		[]
-	);
-	return (
-		<Suspense
-			fallback={
-				<>
-					<h2>{ __( 'Loading…', 'pattern-wrangler' ) }</h2>
-				</>
-			}
-		>
-			<Interface defaults={ defaults } { ...props } />
-		</Suspense>
-	);
-};
 
 const usePatternCategories = ( props ) => {
 	const { getValues } = props;
@@ -85,7 +61,10 @@ const Category = ( props ) => {
 	 * Make sure mapped to category is valid, especially if a mapped category is disabled.
 	 */
 	useEffect( () => {
-		if ( category.mappedTo && ! getCategories().find( ( cat ) => cat.value === category.mappedTo ) ) {
+		if (
+			category.mappedTo &&
+			! getCategories().find( ( cat ) => cat.value === category.mappedTo )
+		) {
 			setValue( `categories.${ category.slug }.mappedTo`, 'none' );
 		}
 	}, [ enabledCategories ] );
@@ -151,8 +130,7 @@ const Category = ( props ) => {
 				</div>
 				<div className="dlx-category-row__label">
 					<div className="dlx-category-row__label-text">
-						{ getLabel() }
-						{ ' ' }
+						{ getLabel() }{ ' ' }
 						<Button
 							variant="link"
 							className="dlx-category-row__label-link"
@@ -162,42 +140,38 @@ const Category = ( props ) => {
 							{ __( 'Edit', 'pattern-wrangler' ) }
 						</Button>
 					</div>
-					<div className="dlx-category-row__slug">
-						{ category.slug }
-					</div>
+					<div className="dlx-category-row__slug">{ category.slug }</div>
 					<div className="dlx-category-row__count">
-						{ category.count } { _n( 'Pattern', 'Patterns', category.count, 'pattern-wrangler' ) }
+						{ category.count }{ ' ' }
+						{ _n( 'Pattern', 'Patterns', category.count, 'pattern-wrangler' ) }
 					</div>
-					{
-						! category.enabled && (
-							<div className="dlx-category-row__map">
-								<Controller
-									name={ `categories.${ category.slug }.mappedTo` }
-									control={ control }
-									render={ ( { field: { onChange, value } } ) => (
-										<SelectControl
-											label={ __( 'Map to Category', 'pattern-wrangler' ) }
-											value={ value }
-											onChange={ ( newValue ) => {
-												onChange( newValue );
-											} }
-											options={ getCategories() }
-										/>
-									) }
-								/>
-							</div>
-						)
-					}
+					{ ! category.enabled && (
+						<div className="dlx-category-row__map">
+							<Controller
+								name={ `categories.${ category.slug }.mappedTo` }
+								control={ control }
+								render={ ( { field: { onChange, value } } ) => (
+									<SelectControl
+										label={ __( 'Map to Category', 'pattern-wrangler' ) }
+										value={ value }
+										onChange={ ( newValue ) => {
+											onChange( newValue );
+										} }
+										options={ getCategories() }
+									/>
+								) }
+							/>
+						</div>
+					) }
 				</div>
 			</div>
 		</>
 	);
 };
 
-const Interface = ( props ) => {
-	const { defaults } = props;
-	const response = defaults();
-	const { data } = response.data;
+const Main = ( props ) => {
+	const data = dlxPatternWranglerAdmin.options;
+	const networkOptions = dlxPatternWranglerAdmin.networkOptions;
 
 	const {
 		control,
@@ -223,6 +197,7 @@ const Interface = ( props ) => {
 			loadCustomizerCSSFrontend: data.loadCustomizerCSSFrontend,
 			hideThemePatterns: data.hideThemePatterns,
 			hidePluginPatterns: data.hidePluginPatterns,
+			enableEnhancedView: data.enableEnhancedView,
 			showMenusUI: data.showMenusUI,
 			categories: data.registered ?? [],
 			makePatternsExportable: data.makePatternsExportable,
@@ -243,192 +218,736 @@ const Interface = ( props ) => {
 				{ Object.values( categories ).map( ( category ) => {
 					return (
 						<li key={ category.slug }>
-							<Category category={ category } control={ control } getValues={ getValues } setValue={ setValue } taxCategories={ data.categories } />
+							<Category
+								category={ category }
+								control={ control }
+								getValues={ getValues }
+								setValue={ setValue }
+								taxCategories={ data.categories }
+							/>
 						</li>
 					);
 				} ) }
 			</ul>
 		);
 	};
+
+	/**
+	 * Check if local patterns can be shown.
+	 *
+	 * @return {boolean} True if local patterns can be shown, false otherwise.
+	 */
+	const canShowLocalPatterns = () => {
+		// if ( dlxPatternWranglerAdmin.isMultisite ) {
+		// 	return (
+		// 		networkOptions.patternConfiguration === 'hybrid' ||
+		// 		networkOptions.patternConfiguration === 'network_only'
+		// 	);
+		// }
+		return true;
+	};
+
+	/**
+	 * Get the show core patterns option.
+	 *
+	 * @return {boolean} True if core patterns should be shown, false otherwise.
+	 */
+	const getShowCoreToggleControl = () => {
+		const corePatternData = {
+			localHidden: false,
+			networkHidden: false,
+		};
+		corePatternData.localHidden = getValues( 'hideCorePatterns' );
+		corePatternData.networkHidden = true;
+		if ( dlxPatternWranglerAdmin.isMultisite ) {
+			if (
+				networkOptions.patternConfiguration === 'hybrid' ||
+				networkOptions.patternConfiguration === 'network_only'
+			) {
+				if ( 'hide' === networkOptions.hideCorePatterns ) {
+					corePatternData.localHidden = true;
+					corePatternData.networkHidden = false;
+				} else if ( 'show' === networkOptions.hideCorePatterns ) {
+					corePatternData.localHidden = false;
+					corePatternData.networkHidden = false;
+				}
+			}
+		}
+		return (
+			<div className="dlx-admin__row">
+				<Controller
+					name="hideCorePatterns"
+					control={ control }
+					render={ ( { field: { onChange, value } } ) => (
+						<ToggleControl
+							label={ __( 'Hide Core Patterns', 'pattern-wrangler' ) }
+							checked={ corePatternData.localHidden }
+							disabled={ ! corePatternData.networkHidden }
+							onChange={ ( boolValue ) => {
+								onChange( boolValue );
+							} }
+							help={ __(
+								'Remove all core patterns from the pattern selector by disabling core patterns.',
+								'pattern-wrangler'
+							) }
+						/>
+					) }
+				/>
+				{ dlxPatternWranglerAdmin.isMultisite &&
+					! corePatternData.networkHidden && (
+					<Notice
+						className="dlx-pw-admin-notice"
+						variant="info"
+						icon={ () => <Info /> }
+					>
+						{ __(
+							'This setting is overridden by the network settings.',
+							'pattern-wrangler'
+						) }
+					</Notice>
+				) }
+			</div>
+		);
+	};
+
+	/**
+	 * Get the show remote patterns option.
+	 *
+	 * @return {React.Component} The remote patterns toggle control.
+	 */
+	const getShowRemotePatternsToggleControl = () => {
+		const remotePatternData = {
+			localHidden: false,
+			networkHidden: false,
+		};
+		remotePatternData.localHidden = getValues( 'hideRemotePatterns' );
+		remotePatternData.networkHidden = true;
+		if ( dlxPatternWranglerAdmin.isMultisite ) {
+			if (
+				networkOptions.patternConfiguration === 'hybrid' ||
+				networkOptions.patternConfiguration === 'network_only'
+			) {
+				if ( 'hide' === networkOptions.hideRemotePatterns ) {
+					remotePatternData.localHidden = true;
+					remotePatternData.networkHidden = false;
+				} else if ( 'show' === networkOptions.hideRemotePatterns ) {
+					remotePatternData.localHidden = false;
+					remotePatternData.networkHidden = false;
+				}
+			}
+		}
+		return (
+			<div className="dlx-admin__row">
+				<Controller
+					name="hideRemotePatterns"
+					control={ control }
+					render={ ( { field: { onChange, value } } ) => (
+						<ToggleControl
+							label={ __( 'Hide Remote Patterns', 'pattern-wrangler' ) }
+							checked={ remotePatternData.localHidden }
+							disabled={ ! remotePatternData.networkHidden }
+							onChange={ ( boolValue ) => {
+								onChange( boolValue );
+							} }
+						/>
+					) }
+				/>
+				{ dlxPatternWranglerAdmin.isMultisite &&
+					! remotePatternData.networkHidden && (
+					<Notice
+						className="dlx-pw-admin-notice"
+						variant="info"
+						icon={ () => <Info /> }
+					>
+						{ __(
+							'This setting is overridden by the network settings.',
+							'pattern-wrangler'
+						) }
+					</Notice>
+				) }
+			</div>
+		);
+	};
+
+	/**
+	 * Get the show unsynced patterns option.
+	 *
+	 * @return {React.Component} The unsynced patterns toggle control.
+	 */
+	const getShowUnsyncedPatternsToggleControl = () => {
+		const unsyncedPatternData = {
+			localHidden: false,
+			networkHidden: false,
+		};
+		unsyncedPatternData.localHidden = getValues( 'hideCoreUnsyncedPatterns' );
+		unsyncedPatternData.networkHidden = true;
+		if ( dlxPatternWranglerAdmin.isMultisite ) {
+			if (
+				networkOptions.patternConfiguration === 'hybrid' ||
+				networkOptions.patternConfiguration === 'network_only'
+			) {
+				if ( 'hide' === networkOptions.hideUnsyncedPatternsForNetwork ) {
+					unsyncedPatternData.localHidden = true;
+					unsyncedPatternData.networkHidden = false;
+				} else if ( 'show' === networkOptions.hideUnsyncedPatternsForNetwork ) {
+					unsyncedPatternData.localHidden = false;
+					unsyncedPatternData.networkHidden = false;
+				}
+			}
+		}
+		return (
+			<div className="dlx-admin__row">
+				<Controller
+					name="hideCoreUnsyncedPatterns"
+					control={ control }
+					render={ ( { field: { onChange, value } } ) => (
+						<ToggleControl
+							label={ __( 'Hide Unsynced Patterns', 'pattern-wrangler' ) }
+							checked={ unsyncedPatternData.localHidden }
+							disabled={ ! unsyncedPatternData.networkHidden }
+							help={ __(
+								'Prevent any unsynced patterns from displaying in the patterns selector. This is useful if you only want to show synced patterns.',
+								'pattern-wrangler'
+							) }
+							onChange={ ( boolValue ) => {
+								onChange( boolValue );
+							} }
+						/>
+					) }
+				/>
+				{ dlxPatternWranglerAdmin.isMultisite &&
+					! unsyncedPatternData.networkHidden && (
+					<Notice
+						className="dlx-pw-admin-notice"
+						variant="info"
+						icon={ () => <Info /> }
+					>
+						{ __(
+							'This setting is overridden by the network settings.',
+							'pattern-wrangler'
+						) }
+					</Notice>
+				) }
+			</div>
+		);
+	};
+
+	/**
+	 * Get the show synced patterns option.
+	 *
+	 * @return {React.Component} The synced patterns toggle control.
+	 */
+	const getShowSyncedPatternsToggleControl = () => {
+		const syncedPatternData = {
+			localHidden: false,
+			networkHidden: false,
+		};
+		syncedPatternData.localHidden = getValues( 'hideCoreSyncedPatterns' );
+		syncedPatternData.networkHidden = true;
+		if ( dlxPatternWranglerAdmin.isMultisite ) {
+			if (
+				networkOptions.patternConfiguration === 'hybrid' ||
+				networkOptions.patternConfiguration === 'network_only'
+			) {
+				if ( 'hide' === networkOptions.hideSyncedPatternsForNetwork ) {
+					syncedPatternData.localHidden = true;
+					syncedPatternData.networkHidden = false;
+				} else if ( 'show' === networkOptions.hideSyncedPatternsForNetwork ) {
+					syncedPatternData.localHidden = false;
+					syncedPatternData.networkHidden = false;
+				}
+			}
+		}
+		return (
+			<div className="dlx-admin__row">
+				<Controller
+					name="hideCoreSyncedPatterns"
+					control={ control }
+					render={ ( { field: { onChange, value } } ) => (
+						<ToggleControl
+							label={ __( 'Hide Synced Patterns', 'pattern-wrangler' ) }
+							checked={ syncedPatternData.localHidden }
+							disabled={ ! syncedPatternData.networkHidden }
+							help={ __(
+								'Prevent any synced patterns from displaying in the patterns selector. This is useful if you only want to show unsynced patterns.',
+								'pattern-wrangler'
+							) }
+							onChange={ ( boolValue ) => {
+								onChange( boolValue );
+							} }
+						/>
+					) }
+				/>
+				{ dlxPatternWranglerAdmin.isMultisite &&
+					! syncedPatternData.networkHidden && (
+					<Notice
+						className="dlx-pw-admin-notice"
+						variant="info"
+						icon={ () => <Info /> }
+					>
+						{ __(
+							'This setting is overridden by the network settings.',
+							'pattern-wrangler'
+						) }
+					</Notice>
+				) }
+			</div>
+		);
+	};
+
+	/**
+	 * Get the show patterns exporter toggle control.
+	 *
+	 * @return {React.Component} The patterns exporter toggle control.
+	 */
+	const getShowPatternsExporterToggleControl = () => {
+		const patternsExporterData = {
+			canExport: getValues( 'makePatternsExportable' ),
+			networkCanExport: true,
+		};
+		if ( dlxPatternWranglerAdmin.isMultisite ) {
+			if (
+				networkOptions.patternConfiguration === 'hybrid' ||
+				networkOptions.patternConfiguration === 'network_only'
+			) {
+				if ( networkOptions.disablePatternExporterForNetwork ) {
+					patternsExporterData.canExport = false;
+					patternsExporterData.networkCanExport = false;
+				} else {
+					patternsExporterData.canExport = getValues( 'makePatternsExportable' );
+					patternsExporterData.networkCanExport = true;
+				}
+			} else {
+				patternsExporterData.canExport = getValues(
+					'disablePatternsExporterBlock'
+				);
+				patternsExporterData.networkCanExport = getValues(
+					'disablePatternsExporterBlock'
+				);
+			}
+		}
+		return (
+			<div className="dlx-admin__row">
+				<Controller
+					name="makePatternsExportable"
+					control={ control }
+					render={ ( { field: { onChange, value } } ) => (
+						<ToggleControl
+							label={ __(
+								'Allow Patterns to be exportable via the WordPress Exporter',
+								'pattern-wrangler'
+							) }
+							checked={ patternsExporterData.canExport }
+							disabled={ ! patternsExporterData.networkCanExport }
+							help={ __(
+								'Enables or disables the default WordPress export feature for content and patterns.',
+								'pattern-wrangler'
+							) }
+							onChange={ ( boolValue ) => {
+								onChange( boolValue );
+							} }
+						/>
+					) }
+				/>
+				{ dlxPatternWranglerAdmin.isMultisite &&
+					! patternsExporterData.networkCanExport && (
+					<Notice
+						className="dlx-pw-admin-notice"
+						variant="info"
+						icon={ () => <Info /> }
+					>
+						{ __(
+							'This setting is overridden by the network settings.',
+							'pattern-wrangler'
+						) }
+					</Notice>
+				) }
+			</div>
+		);
+	};
+
+	/**
+	 * Get the permissions for the pattern importer block.
+	 *
+	 * @return {React.Component} The patterns exporter toggle control.
+	 */
+	const getShowPatternsImporterBlock = () => {
+		const patternsBlockData = {
+			canUseBlock: ! getValues(
+				'disablePatternImporterBlock'
+			),
+			networkCanUseBlock: true,
+		};
+		if ( dlxPatternWranglerAdmin.isMultisite ) {
+			if (
+				networkOptions.patternConfiguration === 'hybrid' ||
+				networkOptions.patternConfiguration === 'network_only'
+			) {
+				if ( networkOptions.disablePatternImporterBlock ) {
+					patternsBlockData.canUseBlock = false;
+					patternsBlockData.networkCanUseBlock = false;
+				} else {
+					patternsBlockData.canUseBlock = true;
+					patternsBlockData.networkCanUseBlock = true;
+				}
+			} else {
+				patternsBlockData.canUseBlock = getValues(
+					'disablePatternImporterBlock'
+				);
+				patternsBlockData.networkCanUseBlock = getValues(
+					'disablePatternImporterBlock'
+				);
+			}
+		}
+		return (
+			<div className="dlx-admin__row">
+				<Controller
+					name="disablePatternImporterBlock"
+					control={ control }
+					render={ ( { field: { onChange, value } } ) => (
+						<ToggleControl
+							label={ __(
+								'Allow Patterns to be imported via the Patterns Importer Block',
+								'pattern-wrangler'
+							) }
+							checked={ false === patternsBlockData.canUseBlock ? false : value }
+							disabled={ ! patternsBlockData.networkCanUseBlock }
+							help={ __(
+								'Disable the patterns importer block, which helps load in remote images.',
+								'pattern-wrangler'
+							) }
+							onChange={ ( boolValue ) => {
+								onChange( boolValue );
+							} }
+						/>
+					) }
+				/>
+				{ dlxPatternWranglerAdmin.isMultisite &&
+					! patternsBlockData.networkCanUseBlock && (
+					<Notice
+						className="dlx-pw-admin-notice"
+						variant="info"
+						icon={ () => <Info /> }
+					>
+						{ __(
+							'This setting is overridden by the network settings.',
+							'pattern-wrangler'
+						) }
+					</Notice>
+				) }
+			</div>
+		);
+	};
+
+	const getShowThemePatternsToggleControl = () => {
+		const themePatternData = {
+			canShow: false,
+			networkCanShow: false,
+		};
+		themePatternData.canShow = getValues( 'hideThemePatterns' );
+		themePatternData.networkCanShow = true;
+		if ( dlxPatternWranglerAdmin.isMultisite ) {
+			if ( 'hide' === networkOptions.hideThemePatterns ) {
+				themePatternData.canShow = false;
+				themePatternData.networkCanShow = false;
+			} else if ( 'show' === networkOptions.hideThemePatterns ) {
+				themePatternData.canShow = true;
+				themePatternData.networkCanShow = false;
+			} else {
+				themePatternData.canShow = getValues( 'hideThemePatterns' );
+				themePatternData.networkCanShow = true;
+			}
+		}
+		return (
+			<div className="dlx-admin__row">
+				<Controller
+					name="hideThemePatterns"
+					control={ control }
+					render={ ( { field: { onChange, value } } ) => (
+						<ToggleControl
+							label={ __(
+								'Hide Theme Patterns',
+								'pattern-wrangler'
+							) }
+							checked={ value || 'default' === value }
+							disabled={ ! themePatternData.networkCanShow }
+							onChange={ ( boolValue ) => {
+								onChange( boolValue );
+							} }
+							help={ __(
+								'Prevent patterns registered by the active theme from displaying in the patterns list.',
+								'pattern-wrangler'
+							) }
+						/>
+					) }
+				/>
+				{ dlxPatternWranglerAdmin.isMultisite &&
+					! themePatternData.networkCanShow && (
+					<Notice
+						className="dlx-pw-admin-notice"
+						variant="info"
+						icon={ () => <Info /> }
+					>
+						{ __(
+							'This setting is overridden by the network settings.',
+							'pattern-wrangler'
+						) }
+					</Notice>
+				) }
+			</div>
+		);
+	};
+
+	const getShowPluginPatternsToggleControl = () => {
+		const pluginPatternData = {
+			canShow: false,
+			networkCanShow: false,
+		};
+		pluginPatternData.canShow = getValues( 'hidePluginPatterns' );
+		pluginPatternData.networkCanShow = true;
+		if ( dlxPatternWranglerAdmin.isMultisite ) {
+			if ( 'hide' === networkOptions.hidePluginPatterns ) {
+				pluginPatternData.canShow = false;
+				pluginPatternData.networkCanShow = false;
+			} else if ( 'show' === networkOptions.hidePluginPatterns ) {
+				pluginPatternData.canShow = true;
+				pluginPatternData.networkCanShow = false;
+			}
+		}
+		return (
+			<div className="dlx-admin__row">
+				<Controller
+					name="hidePluginPatterns"
+					control={ control }
+					render={ ( { field: { onChange, value } } ) => (
+						<ToggleControl
+							label={ __(
+								'Hide Plugin Patterns',
+								'pattern-wrangler'
+							) }
+							checked={ value || 'default' === 'value' }
+							disabled={ ! pluginPatternData.networkCanShow }
+							onChange={ ( boolValue ) => {
+								onChange( boolValue );
+							} }
+							help={ __(
+								'Prevent patterns registered by active plugins from displaying in the patterns list.',
+								'pattern-wrangler'
+							) }
+						/>
+					) }
+				/>
+				{ dlxPatternWranglerAdmin.isMultisite &&
+					! pluginPatternData.networkCanShow && (
+					<Notice
+						className="dlx-pw-admin-notice"
+						variant="info"
+						icon={ () => <Info /> }
+					>
+						{ __(
+							'This setting is overridden by the network settings.',
+							'pattern-wrangler'
+						) }
+					</Notice>
+				) }
+			</div>
+		);
+	};
+
+	/**
+	 * Get the Hide All Patterns togggle control.
+	 *
+	 * @return {React.Component} The patterns exporter toggle control.
+	 */
+	const getHideAllPatternsToggleControl = () => {
+		const hideAllPatternsData = {
+			allPatternsDisabled: false,
+			networkAllPatternsDisabled: false,
+		};
+		if ( dlxPatternWranglerAdmin.isMultisite ) {
+			if ( networkOptions.patternConfiguration === 'disabled' || 'hide' === networkOptions.hideAllPatterns ) {
+				hideAllPatternsData.allPatternsDisabled = true;
+				hideAllPatternsData.networkAllPatternsDisabled = true;
+			} else {
+				hideAllPatternsData.allPatternsDisabled = getValues( 'hideAllPatterns' );
+				hideAllPatternsData.networkAllPatternsDisabled = false;
+			}
+		} else {
+			hideAllPatternsData.allPatternsDisabled = getValues( 'hideAllPatterns' );
+			hideAllPatternsData.networkAllPatternsDisabled = false;
+		}
+		return (
+			<div className="dlx-admin__row">
+				<Controller
+					name="hideAllPatterns"
+					control={ control }
+					render={ ( { field: { onChange, value } } ) => (
+						<ToggleControl
+							label={ __( 'Hide All Patterns', 'pattern-wrangler' ) }
+							checked={ hideAllPatternsData.allPatternsDisabled }
+							disabled={ hideAllPatternsData.networkAllPatternsDisabled }
+							help={ __(
+								'Disable all patterns and the pattern selector.',
+								'pattern-wrangler'
+							) }
+							onChange={ ( boolValue ) => {
+								onChange( boolValue );
+							} }
+						/>
+					) }
+				/>
+				{ dlxPatternWranglerAdmin.isMultisite &&
+					hideAllPatternsData.networkAllPatternsDisabled && (
+					<Notice
+						className="dlx-pw-admin-notice"
+						variant="info"
+						icon={ () => <Info /> }
+					>
+						{ __(
+							'This setting is overridden by the network settings.',
+							'pattern-wrangler'
+						) }
+					</Notice>
+				) }
+			</div>
+		);
+	};
+
 	return (
 		<>
 			<div className="dlx-pw-admin-content-heading">
-				<h1><span className="dlx-pw-content-heading-text">{ __( 'Settings for Pattern Wrangler', 'pattern-wrangler' ) }</span></h1>
+				<h1>
+					<span className="dlx-pw-content-heading-text">
+						{ __( 'Settings for Pattern Wrangler', 'pattern-wrangler' ) }
+					</span>
+				</h1>
 				<p className="description">
-					{
-						__( 'Configure which patterns are displayed and adjust settings and categories.', 'pattern-wrangler' )
-					}
+					{ __(
+						'Configure which patterns are displayed and adjust settings and categories.',
+						'pattern-wrangler'
+					) }
 				</p>
+				{ dlxPatternWranglerAdmin.isMultisite &&
+					dlxPatternWranglerAdmin.isUserNetworkAdmin && (
+					<Notice
+						className="dlx-pw-admin-notice"
+						variant="info"
+						icon={ () => <Info /> }
+					>
+						<div>
+							{ __(
+								'This is a multisite installation. You can manage network settings by clicking the button below.',
+								'pattern-wrangler'
+							) }
+						</div>
+						<div>
+							<Button
+								variant="link"
+								href={ dlxPatternWranglerAdmin.networkAdminSettingsUrl }
+								target="_blank"
+							>
+								{ __( 'Network Settings', 'pattern-wrangler' ) }
+							</Button>
+						</div>
+					</Notice>
+				) }
 			</div>
 			{ /* eslint-disable-next-line no-unused-vars */ }
-			<form onSubmit={ handleSubmit( ( formData ) => { } ) }>
+			<form onSubmit={ handleSubmit( ( formData ) => {} ) }>
 				<div id="dlx-pw-admin-table">
 					<table className="form-table form-table-row-sections">
 						<tbody>
+							<tr>
+								<th scope="row">{ __( 'Enhanced View', 'pattern-wrangler' ) }</th>
+								<td>
+									<Controller
+										name="enableEnhancedView"
+										control={ control }
+										render={ ( { field: { onChange, value } } ) => (
+											<ToggleControl
+												label={ __(
+													'Enable Enhanced Patterns View',
+													'pattern-wrangler'
+												) }
+												checked={ value }
+												onChange={ ( boolValue ) => {
+													onChange( boolValue );
+												} }
+												help={ __(
+													'This will enable the enhanced patterns view when viewing all patterns rather than showing a classic interface.',
+													'pattern-wrangler'
+												) }
+											/>
+										) }
+									/>
+								</td>
+							</tr>
 							<tr>
 								<th scope="row">
 									{ __( 'Pattern Visibility', 'pattern-wrangler' ) }
 								</th>
 								<td>
-									<div className="dlx-admin__row">
-										<Controller
-											name="hideAllPatterns"
-											control={ control }
-											render={ ( { field: { onChange, value } } ) => (
-												<div className="dlx-admin__tooltip-toggle">
-													<Tooltip
-														text={ __( 'This option is useful if you want to hide all patterns from the block editor.', 'pattern-wrangler' ) }
-													>
+									{ getHideAllPatternsToggleControl() }
+									{ canShowLocalPatterns() && (
+										<>
+											{ getValues( 'hideAllPatterns' ) && (
+												<div className="dlx-admin__row">
+													<Controller
+														name="hidePatternsMenu"
+														control={ control }
+														render={ ( { field: { onChange, value } } ) => (
+															<ToggleControl
+																label={ __(
+																	'Hide Pattern Wrangler Menu Item',
+																	'pattern-wrangler'
+																) }
+																checked={ value }
+																onChange={ ( boolValue ) => {
+																	onChange( boolValue );
+																} }
+																help={ __(
+																	'This will disable the top-level menu and move the Patterns menu under Appearance.',
+																	'pattern-wrangler'
+																) }
+															/>
+														) }
+													/>
+												</div>
+											) }
+											{ getShowCoreToggleControl() }
+											{ getShowRemotePatternsToggleControl() }
+											{ getShowThemePatternsToggleControl() }
+											{ getShowPluginPatternsToggleControl() }
+											{ getShowUnsyncedPatternsToggleControl() }
+											{ getShowSyncedPatternsToggleControl() }
+											<div className="dlx-admin__row">
+												<Controller
+													name="hideUncategorizedPatterns"
+													control={ control }
+													render={ ( { field: { onChange, value } } ) => (
 														<ToggleControl
-															label={ __( 'Hide All Patterns', 'pattern-wrangler' ) }
+															label={ __(
+																'Hide Uncategorized Patterns',
+																'pattern-wrangler'
+															) }
 															checked={ value }
 															onChange={ ( boolValue ) => {
 																onChange( boolValue );
 															} }
-															help={ __( 'Disable all patterns and the pattern selector.', 'pattern-wrangler' ) }
-															tooltip={ __( 'This option is useful if you want to hide all patterns from the block editor.', 'pattern-wrangler' ) }
+															help={ __(
+																'Prevent any patterns not in any registered categories from displaying.',
+																'pattern-wrangler'
+															) }
 														/>
-													</Tooltip>
-												</div>
-											) }
-										/>
-									</div>
-									{ getValues( 'hideAllPatterns' ) && (
-										<div className="dlx-admin__row">
-											<Controller
-												name="hidePatternsMenu"
-												control={ control }
-												render={ ( { field: { onChange, value } } ) => (
-													<ToggleControl
-														label={ __( 'Hide Pattern Wrangler Menu Item', 'pattern-wrangler' ) }
-														checked={ value }
-														onChange={ ( boolValue ) => {
-															onChange( boolValue );
-														} }
-														help={ __( 'This will disable the top-level menu and move the Patterns menu under Appearance.', 'pattern-wrangler' ) }
-													/>
-												) }
-											/>
-										</div>
+													) }
+												/>
+											</div>
+										</>
 									) }
-									<div className="dlx-admin__row">
-										<Controller
-											name="hideCorePatterns"
-											control={ control }
-											render={ ( { field: { onChange, value } } ) => (
-												<ToggleControl
-													label={ __( 'Hide Core Patterns', 'pattern-wrangler' ) }
-													checked={ value }
-													onChange={ ( boolValue ) => {
-														onChange( boolValue );
-													} }
-													help={ __( 'Remove all core patterns from the pattern selector by disabling core patterns.', 'pattern-wrangler' ) }
-												/>
-											) }
-										/>
-									</div>
-									<div className="dlx-admin__row">
-										<Controller
-											name="hideRemotePatterns"
-											control={ control }
-											render={ ( { field: { onChange, value } } ) => (
-												<ToggleControl
-													label={ __( 'Hide Remote Patterns', 'pattern-wrangler' ) }
-													checked={ value }
-													onChange={ ( boolValue ) => {
-														onChange( boolValue );
-													} }
-													help={ __( 'Prevent users from searching for remote patterns in the pattern selector.', 'pattern-wrangler' ) }
-												/>
-											) }
-										/>
-									</div>
-									<div className="dlx-admin__row">
-										<Controller
-											name="hideThemePatterns"
-											control={ control }
-											render={ ( { field: { onChange, value } } ) => (
-												<ToggleControl
-													label={ __( 'Hide Theme Patterns', 'pattern-wrangler' ) }
-													checked={ value }
-													onChange={ ( boolValue ) => {
-														onChange( boolValue );
-													} }
-													help={ __( 'Prevent patterns registered by the active theme from displaying in the patterns list.', 'pattern-wrangler' ) }
-												/>
-											) }
-										/>
-									</div>
-									<div className="dlx-admin__row">
-										<Controller
-											name="hidePluginPatterns"
-											control={ control }
-											render={ ( { field: { onChange, value } } ) => (
-												<ToggleControl
-													label={ __( 'Hide Plugin Patterns', 'pattern-wrangler' ) }
-													checked={ value }
-													onChange={ ( boolValue ) => {
-														onChange( boolValue );
-													} }
-													help={ __( 'Prevent patterns registered by active plugins from displaying in the patterns list.', 'pattern-wrangler' ) }
-												/>
-											) }
-										/>
-									</div>
-									<div className="dlx-admin__row">
-										<Controller
-											name="hideCoreSyncedPatterns"
-											control={ control }
-											render={ ( { field: { onChange, value } } ) => (
-												<ToggleControl
-													label={ __( 'Hide Synced Patterns', 'pattern-wrangler' ) }
-													checked={ value }
-													onChange={ ( boolValue ) => {
-														onChange( boolValue );
-													} }
-													help={ __( 'Prevent any synced patterns (reusable blocks) from displaying in the patterns selector. This will prevent users from using synced patterns in the block editor.', 'pattern-wrangler' ) }
-												/>
-											) }
-										/>
-									</div>
-									<div className="dlx-admin__row">
-										<Controller
-											name="hideCoreUnsyncedPatterns"
-											control={ control }
-											render={ ( { field: { onChange, value } } ) => (
-												<ToggleControl
-													label={ __( 'Hide Unsynced Patterns', 'pattern-wrangler' ) }
-													checked={ value }
-													onChange={ ( boolValue ) => {
-														onChange( boolValue );
-													} }
-													help={ __( 'Prevent any unsynced patterns from displaying in the patterns selector. This is useful if you only want to show synced patterns.', 'pattern-wrangler' ) }
-												/>
-											) }
-										/>
-									</div>
-									<div className="dlx-admin__row">
-										<Controller
-											name="hideUncategorizedPatterns"
-											control={ control }
-											render={ ( { field: { onChange, value } } ) => (
-												<ToggleControl
-													label={ __( 'Hide Uncategorized Patterns', 'pattern-wrangler' ) }
-													checked={ value }
-													onChange={ ( boolValue ) => {
-														onChange( boolValue );
-													} }
-													help={ __( 'Prevent any patterns not in any registered categories from displaying.', 'pattern-wrangler' ) }
-												/>
-											) }
-										/>
-									</div>
 								</td>
 							</tr>
 							<tr>
-								<th scope="row">
-									{ __( 'Customizer', 'pattern-wrangler' ) }
-								</th>
+								<th scope="row">{ __( 'Customizer', 'pattern-wrangler' ) }</th>
 								<td>
 									<div className="dlx-admin__row">
 										<Controller
@@ -441,7 +960,10 @@ const Interface = ( props ) => {
 													onChange={ ( boolValue ) => {
 														onChange( boolValue );
 													} }
-													help={ __( 'This will show the customizer UI in the Appearance menu if enabled.', 'pattern-wrangler' ) }
+													help={ __(
+														'This will show the customizer UI in the Appearance menu if enabled.',
+														'pattern-wrangler'
+													) }
 												/>
 											) }
 										/>
@@ -452,12 +974,18 @@ const Interface = ( props ) => {
 											control={ control }
 											render={ ( { field: { onChange, value } } ) => (
 												<ToggleControl
-													label={ __( 'Load Customizer CSS in the Block Editor', 'pattern-wrangler' ) }
+													label={ __(
+														'Load Customizer CSS in the Block Editor',
+														'pattern-wrangler'
+													) }
 													checked={ value }
 													onChange={ ( boolValue ) => {
 														onChange( boolValue );
 													} }
-													help={ __( 'This will load any CSS in the customizer in the block editor as well.', 'pattern-wrangler' ) }
+													help={ __(
+														'This will load any CSS in the customizer in the block editor as well.',
+														'pattern-wrangler'
+													) }
 												/>
 											) }
 										/>
@@ -468,12 +996,18 @@ const Interface = ( props ) => {
 											control={ control }
 											render={ ( { field: { onChange, value } } ) => (
 												<ToggleControl
-													label={ __( 'Load Customizer CSS on the Frontend', 'pattern-wrangler' ) }
+													label={ __(
+														'Load Customizer CSS on the Frontend',
+														'pattern-wrangler'
+													) }
 													checked={ value }
 													onChange={ ( boolValue ) => {
 														onChange( boolValue );
 													} }
-													help={ __( 'By default, WordPress loads customizer CSS on the frontend. Disable this option to prevent any customizer CSS from loading.', 'pattern-wrangler' ) }
+													help={ __(
+														'By default, WordPress loads customizer CSS on the frontend. Disable this option to prevent any customizer CSS from loading.',
+														'pattern-wrangler'
+													) }
 												/>
 											) }
 										/>
@@ -481,9 +1015,7 @@ const Interface = ( props ) => {
 								</td>
 							</tr>
 							<tr>
-								<th scope="row">
-									{ __( 'Miscellaneous', 'pattern-wrangler' ) }
-								</th>
+								<th scope="row">{ __( 'Miscellaneous', 'pattern-wrangler' ) }</th>
 								<td>
 									<div className="dlx-admin__row">
 										<Controller
@@ -496,59 +1028,38 @@ const Interface = ( props ) => {
 													onChange={ ( boolValue ) => {
 														onChange( boolValue );
 													} }
-													help={ __( 'This will show the menus UI in the Appearance menu if enabled.', 'pattern-wrangler' ) }
+													help={ __(
+														'This will show the menus UI in the Appearance menu if enabled.',
+														'pattern-wrangler'
+													) }
 												/>
 											) }
 										/>
 									</div>
-									<div className="dlx-admin__row">
-										<Controller
-											name="disablePatternImporterBlock"
-											control={ control }
-											render={ ( { field: { onChange, value } } ) => (
-												<ToggleControl
-													label={ __( 'Disable Patterns Importer Block', 'pattern-wrangler' ) }
-													checked={ value }
-													onChange={ ( boolValue ) => {
-														onChange( boolValue );
-													} }
-													help={ __( 'Disable the pattern importer block, which helps load in remote images.', 'pattern-wrangler' ) }
-												/>
-											) }
-										/>
-									</div>
+									{ getShowPatternsImporterBlock() }
 									<div className="dlx-admin__row">
 										<Controller
 											name="allowFrontendPatternPreview"
 											control={ control }
 											render={ ( { field: { onChange, value } } ) => (
 												<ToggleControl
-													label={ __( 'Enable a Pattern Preview on the Frontend', 'pattern-wrangler' ) }
+													label={ __(
+														'Enable a Pattern Preview on the Frontend',
+														'pattern-wrangler'
+													) }
 													checked={ value }
 													onChange={ ( boolValue ) => {
 														onChange( boolValue );
 													} }
-													help={ __( 'This will enable previews in the patterns post type so you can preview a pattern as if it were on a page.', 'pattern-wrangler' ) }
+													help={ __(
+														'This will enable previews in the patterns post type so you can preview a pattern as if it were on a page.',
+														'pattern-wrangler'
+													) }
 												/>
 											) }
 										/>
 									</div>
-									<div className="dlx-admin__row">
-										<Controller
-											name="makePatternsExportable"
-											control={ control }
-											render={ ( { field: { onChange, value } } ) => (
-												<ToggleControl
-													label={ __( 'Allow Patterns to be exportable via the WordPress Exporter', 'pattern-wrangler' ) }
-													checked={ value }
-													onChange={ ( boolValue ) => {
-														onChange( boolValue );
-													} }
-													help={ __( 'This will allow Patterns to be exportable. When you are done exporting, it is advised to switch this back off.', 'pattern-wrangler' ) }
-												/>
-											) }
-										/>
-									</div>
+									{ getShowPatternsExporterToggleControl() }
 								</td>
 							</tr>
 							<tr className="dlx-table-row-categories">
@@ -556,15 +1067,18 @@ const Interface = ( props ) => {
 									{ __( 'Pattern Categories', 'pattern-wrangler' ) }
 								</th>
 								<td>
-									{ ( Object.values( getValues( 'categories' ) ).length === 0 ) && (
-										<div className="dlx-admin__row dlx-admin__row-full-width" >
+									{ Object.values( getValues( 'categories' ) ).length === 0 && (
+										<div className="dlx-admin__row dlx-admin__row-full-width">
 											<p>
-												{ __( 'No categories have been registered via core, themes or plugins.', 'pattern-wrangler' ) }
+												{ __(
+													'No categories have been registered via core, themes or plugins.',
+													'pattern-wrangler'
+												) }
 											</p>
 										</div>
 									) }
-									{ ( Object.values( getValues( 'categories' ) ).length > 0 ) && (
-										<div className="dlx-admin__row dlx-admin__row-full-width" >
+									{ Object.values( getValues( 'categories' ) ).length > 0 && (
+										<div className="dlx-admin__row dlx-admin__row-full-width">
 											<PanelBody
 												title={ __( 'Pattern Categories', 'pattern-wrangler' ) }
 												initialOpen={ false }
